Update functors to work with ParquetTable objects, and add tests
Before, the __call__ method of a `Functor` expected to take a `Catalog` as argument; this changes things so that it takes a `ParquetTable` object. One of the important things that change here is a functor now specifies a "filt" and "dataset" that it acts upon.  This can be specified upon declaration of a Functor, or not specified, in which case default values are used (defaults specified in functor definition).
This commit also includes changes to how `from_yaml` parses a yaml functor specifying a `CompositeFunctor`, so that it parses abitrary kwargs, and takes a 'flags' section without specifying what dataset to be taken from (in which case the flags are taken from the 'ref' dataset).
In addition, this makes dropping levels optional when calling the
`.toDataFrame` method of `MultilevelParquetTable`.  This is necessary
due to the new behavior of the functors, in particular CompositeFunctor.