Implement PostprocessAnalysis and PostprocessTask objects; remove writeQATable.py
The `PostprocessAnalysis` object takes over the calculation-of-columns-from-catalog functionality that used to be within QADataset.  This is a generalized object that takes a `ParquetTable` object and a list/dict of functors (or a CompositeFunctor).  After initialization, accessing the `.df` attribute triggers computation.  `PostprocessTask` is a general purpose task that uses the PostprocessAnalysis object to compute a new output dataset given a YAML file that specifies a CompositeFunctor (`--functorFile` config parameter).  `MultibandPostprocessTask` computes the requested set of functors for each filter.  This completely replaces `WriteQATableTask` and the old `PostprocessCatalogTask`, which were defined in `writeQATable.py`, which is removed by this commit.