Use rasterize to enable colorbar for sky plots.
The `holoviews.operations.datashader.datashade` operation does not allow for a colorbar; however, `rasterize` does.  This changes the "sky" part of the "scattersky" plots to use rasterize and thus enable a colorbar.  The downside to this is that some of the functionality of 'datashade' goes away, such as the ability to use `dynspread` to increase the size of "points" for clarity when zoomed in.