add support for noDups and flattening multilevel result
This gives the option to have the output of a MultibandPostprocessingTask
to be a flat dataframe (as opposed to a df with multi-level column
index).  If this is the case, then the columns will be named with a
filter prefix (as defined by .filterMap), except for those functors
that are defined to be noDup=True.

made default noDups

make sure columns are not duplicated

fix flattening if filter is not there