Fix for removal of MergeSourcesTask
The thin baseclass MergeSourcesTask was recently removed, as it was
only providing a mix of related method, that each derived class only
used a fraction of and overloaded a lot of the remaining ones. This
complicated many of the derived class implementations. Much of the
shared code was separated out into free functions were appropriate.

This class has been updated to use these free functions, and other
trivial methods were included directly in its implementation.